+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.11.0
 
 -   Fixed: `rateLimit` should remove an item if it has rejected.
+-   `rateLimit` - `.finish` swallows exceptions.
 -   `rateLimit` - Added `length` getter.
 -   `batchQueue` - Added `length` getter.
 -   `throughputLimiter.getCurrentRate` - always returns an integer.